adding components into place into main app
