+++ src/components/App.jsx
@@ -3,10 +3,10 @@ import Example from "../containers/exampleCounterContainer.js";
 
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-import LandingPage from "./MainComponents/LandingPage_Components/LandingPage.jsx";
-import Map from "./MainComponents/Map_Components/Map.jsx";
+import Banner from "./MainComponents/Banner/Banner.jsx";
+import LandingPage from "./MainComponents/LandingPage_Components/LandingPage";
+import Dashboard from './MainComponents/Trip-Planner/Dashboard';
 import Itinerary from "./MainComponents/Itinerary_Components/Itinerary.jsx";
-import Banner from "./Banner/Banner.jsx";
 
 const App = ({ name }) => {
   return (
@@ -18,13 +18,12 @@ const App = ({ name }) => {
         <Switch>
           <Route path="/" exact>
             <Banner />
-            <LandingPage exampleProps={{ example: true }} />
-            {/* ^ this is v5 react router way to pass props without a cb */}
+            <LandingPage />
           </Route>
 
           <Route path="/map" exact>
-            <Banner />
-            <Map exampleProps={{ example: true }} />
+            {/* <Banner /> */}
+            <Dashboard />
           </Route>
 
           <Route path="/itinerary/:id" exact>
