--- src/components/MainComponents/Map_Components/Map.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Map = () => {
-    return <div> This is where the Map would go (parent component) </div>
-}
-
-export default Map
