+++ src/components/MainComponents/Trip-Planner/Suggestions/Filter-Slider.js
