+++ src/index.js
@@ -4,7 +4,7 @@ import App from "./components/App.jsx";
 import Dashboard from "./components/Trip-Planner/Dashboard.js";
 import { Provider } from "react-redux";
 import { store } from "./store.js";
-import LandingPage from './components/LandingPage'
+import LandingPage from './components/MainComponents/LandingPage_Components/LandingPage'
 ReactDOM.render(
 	<Provider store={store}>
 		<LandingPage />
